+++ src/images_upload_cli/util.py
@@ -6,8 +6,12 @@ from shutil import which
 from subprocess import Popen
 
 import click
+from httpx import Response
+from loguru import logger
 
-from images_upload_cli.error import GetEnvError
+
+class GetEnvError(Exception):
+    """Exception raised when an environment variable is not found."""
 
 
 def get_config_path() -> Path:
@@ -75,3 +79,28 @@ def notify_send(text_to_print: str) -> None:
     """
     if notify_send := which("notify-send"):
         Popen([notify_send, "-a", "images-upload-cli", text_to_print])
+
+
+def log_on_error(response: Response) -> None:
+    """
+    Logs an error message based on the HTTP response.
+
+    Args:
+        response (Response): The HTTP response object.
+
+    Returns:
+        None
+    """
+    status_class = response.status_code // 100
+    error_types = {
+        1: "Informational response",
+        3: "Redirect response",
+        4: "Client error",
+        5: "Server error",
+    }
+    error_type = error_types.get(status_class, "Invalid status code")
+
+    logger.error(
+        f"{error_type} '{response.status_code} {response.reason_phrase}' for url '{response.url}'."
+    )
+    logger.debug(f"Response text:\n{response.text}")
